% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sero_Placement_Functions.R
\name{Clade_Hit_Finder_Pendant_Final}
\alias{Clade_Hit_Finder_Pendant_Final}
\title{MRCA Calculation From Placement Data}
\usage{
Clade_Hit_Finder_Pendant_Final(Pendant_Multi, Tree)
}
\arguments{
\item{Pendant_Multi}{An integer which is the multiplier by which to scale the pendant length adjustment in step 4.}

\item{Tree}{A phylogenetic tree representing the reference sequences in Newick format. This tree is provided as
part of the Seroplacer package.}
}
\value{
An integer which is the node value in Tree of the final pendant-adjusted MRCA.
}
\description{
This function is the final optimized version of the clade finding algorithm which we apply to a set of placement
results. The steps for this algorithm are described below:
1. EPA-NG is ran given the query sequence and all reported edges within the 99.9% likelihood weight ratio (LWR)
range are considered for possible initial hits. Query sequences are inserted in between two nodes by EPA-NG, and
either the distal or proximal node is chosen depending on which is closer to the query insertion point.
2. All tips which are descendants of the closest node to the query branch insertion point are considered to be
the set of initial hits.
3. The most recent common ancestor (MRCA) of these initial hits is calculated and considered the initial clade MRCA.
4. The maximum pendant length from the reported edges in EPA-NG is recorded and multiplied by a scaler in order
to perform the pendant length adjustment. The scaler is varied in order to optimize for best algorithm performance.
5. Pendant length adjustment is done by moving a pairwise distance from the initial clade MRCA towards the tree
root by a multiple of the maximum pendant length value. After traveling up the tree to a new position, the closest
node (distal or proximal) to this position is recorded and labeled as the pendant adjusted MRCA. If the pendant
length is small enough, it is possible that the MRCA will not change during pendant length adjustment.
6. All descendants of this pendant adjusted MRCA are considered the final results of the placement.
}
\examples{
MRCA <- Clade_Hit_Finder_Pendant_Final(Pendant_Multi = 1.5, Tree = full.tree)
}
