% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sero_Placement_Functions.R
\name{Concatenate_By_Order}
\alias{Concatenate_By_Order}
\title{Reference Sequence Concatenation}
\usage{
Concatenate_By_Order(Order, RefSequence)
}
\arguments{
\item{Order}{A vector containing the integers 1-7 in an order which the reference alleles should be sorted into for
matching with the query sequence.}

\item{RefSequence}{A vector of nucleotide sequences containing 7 sequences representing the 7 16s reference
alleles. This vector should contain 8 total elements, the first being a name, followed by 7 sequence elements.}
}
\value{
A character string containing 7 concatenated 16S sequences, in the order provided by the order parameter.
}
\description{
EPA-NG placement input requires a single sequence, so we concatenate the multi-copy alleles to resolve this. This
function concatenates a set of reference alleles based on the previously calculated best order determined using
Super_HAM_Placer.
}
\examples{
Concatenated_Sequences <- Concatenate_By_Order(Order = c(1,4,2,3,6,7,5), RefSequence = Reference_Sequence)
}
