% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sero_Placement_Functions.R
\name{Data_Preparation}
\alias{Data_Preparation}
\title{Query and Reference Sequence Data Preparation}
\usage{
Data_Preparation(PathToAlignedQuery, Aligned_ReferenceFile)
}
\arguments{
\item{PathToAlignedQuery}{A file path leading to the directory containing the pre-aligned query sequences. This
file should contain a set of 7 ASVs in a FASTA format file which have been aligned to the reference sequence
database using MAFFT.}

\item{Aligned_ReferenceFile}{(Provided in package) Select the correct 16S sequence reference file for the
organism being studied (Salmonella or E. coli).}
}
\value{
A list with 2 elements containing the user's formatted sequences. List element 1 contains the concatenated
query sequences. List element 2 contains the re-sorted and concatenated reference sequence file. Both of these
will be used as inputs for the EPA-NG wrapper function.
}
\description{
Takes a pre-aligned set of query sequences from the user and formats the query and reference sequence file into
the appropriate formats for EPA-NG. This function executes the initial step of optimizing the order of the 7
16S reference sequences into the best possible orientation matching the user's query. The best orientation for
each reference is calculated and then the reference sequences are re-sorted into the optimal orders. After this,
the reference sequences are all concatenated and returned in a format ready for EPA-NG. In addition, the user's
input query sequences are also concatenated and returned as a single FASTA line for input into EPA-NG. Note: It
is required that the input query sequences are pre-aligned to the main reference file using MAFFT. Non-aligned
sequences with varying sequence lengths will cause errors in downstream analysis steps.
}
\examples{
Sequence_Outputs<-Data_Preparation(PathToAlignedQuery = "/path/to/your/file.fasta",Aligned_ReferenceFile = Full_16s_Data_Aligned)
}
