% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sero_Placement_Functions.R
\name{Depth_Calculator}
\alias{Depth_Calculator}
\title{Maximum Depth and Sum of Branch Length Calculation}
\usage{
Depth_Calculator(Serovar, Dists, Serovar_Names)
}
\arguments{
\item{Serovar}{A character string containing a serovar name.}

\item{Dists}{A subset of the table of pairwise distances between all pairs of nodes in the phylogenetic tree.
Calculated from dist.nodes {ape}. This subset of pairwise distances should contain specifically the distances
between the MRCA and its descendants.}

\item{Serovar_Names}{A character vector of all serovar names in the set of resulting hits.}
}
\value{
A numeric vector containing the maximum depth and sum of branch lengths measured in pairwise distance.
}
\description{
This function is built into the Placement_Results_Output function which generates serovar placement results data.
This function calculates the maximum depth, or the distance from the MRCA to the farthest hit in the clade, and also
the sum of branch lengths, or the total sum of branch lengths from each hit to the MRCA.
}
\examples{
Missing
}
