% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sero_Placement_Functions.R
\name{mafft_wrap}
\alias{mafft_wrap}
\title{Wrapper Function for MAFFT}
\usage{
mafft_wrap(x, y, exec, options, file, add)
}
\arguments{
\item{x}{A FASTA file containing the appropriate organism's multiple sequence alignment. These are pre-built and
provided with the package, and the correct one must be selected. The user's query will be aligned matching this
multiple sequence alignment.}

\item{y}{A set of amplicon sequencing variants (ASVs) representing the 7 16S alleles. These are the query sequences
which will be aligned to the multiple sequence alignment in x.}

\item{exec}{(Optional) The path on the machine in which the MAFFT program is installed. This path will default to
the default installation path if left blank (/usr/local/bin/mafft).}

\item{options}{(Optional) Option settings which can be passed to the MAFFT command. The default setting will add the
option --keeplength. This option is required for this analysis, as it will ensure that the length of the sequences
in the multiple sequence alignment is not changed after the addition of the user's query sequences.}

\item{file}{(Optional) A path to which the output FASTA should be saved. If left blank, the wrapper function will
output the resulting FASTA as an R environment object.}

\item{add}{Attaches the --add argument to the MAFFT command, setting the program to append the user's query sequences
to the provided reference multiple sequence alignment. This command is required for merging the query sequences into
the reference FASTA file.}
}
\value{
A multiple sequence alignment in FASTA format containing the reference sequences and the aligned query
sequences appended as the last sequences in the file. The sequences in this alignment will be identical in length
to the original reference multiple sequence alignment.
}
\description{
Creates an R interface for using the command line tool MAFFT in an R environment. It is required that MAFFT is
initially installed on the machine. Installation instructions for MAC OS X and Windows can be located at
<https://mafft.cbrc.jp/alignment/software/source.html>. The installation path should be noted and may be needed
as a variable in the wrapper function. This wrapper is a more specific implementation of MAFFT which will take
the user's input query sequences and align them to a previously aligned reference sequence database.
The input sequences will be aligned with their lengths adjusted to match the reference multiple sequence
alignment (MSA). The input sequences will then be appended to the end of the MSA for use in downstream analysis.
This wrapper is not currently designed to include all options contained in MAFFT since our implementation requires
a specific set of parameters to be used.
}
\examples{
Output_FASTA<-mafft_wrap(x = Full_Alignment.fasta, y = Query.fasta, options = "--keeplength")
}
