% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sero_Placement_Functions.R
\name{MSA_Trim}
\alias{MSA_Trim}
\title{Trimming Query Sequences from MAFFT}
\usage{
MSA_Trim(MSA, File)
}
\arguments{
\item{MSA}{A multiple sequence alignment file outputted from the MAFFT alignment step. This MSA should contain
a set of reference sequences with 7 appended query sequences at the end which will be extracted and reformatted.}

\item{File}{(Optional) A path to which to save the aligned query sequences. If a path is provided, the queries
will be saved as a .Rdata object with a data frame containing the query sequences.}
}
\value{
A data frame with 2 columns and 7 rows containing the names of the query sequences in the first column
and the nucleotide sequences in the second column. These 7 sequences are the aligned queries submitted in the
first MAFFT step of the protocol.
}
\description{
This function will extract the aligned query sequences from the MAFFT alignment file. After initially aligning
the query sequences to the pre-aligned reference sequence database, the last 7 sequences in the output FASTA
file will be the newly aligned query sequences. This function will extract those 7 sequences and format them
into the appropriate format for the Data_Preparation function. It is important to run this function as the Data
Preparation function requires the sequences to be in a data frame, and not a DNAbin object or other format.
}
\examples{
Aligned_Query_FASTA <- MSA_Trim(Output_FASTA)
}
