% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sero_Placement_Functions.R
\name{no_legend}
\alias{no_legend}
\title{Legend Removal for Plotting}
\usage{
no_legend()
}
\value{
N/A
}
\description{
This function removes the legend from the phylogeny plot for visualization.
}
\examples{
None
}
