% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sero_Placement_Functions.R
\name{Phylogeny_Plotting}
\alias{Phylogeny_Plotting}
\title{Phylogeny Plotting}
\usage{
Phylogeny_Plotting(MRCA, Species)
}
\arguments{
\item{MRCA}{An integer value for the node which is the final pendant adjusted MRCA.}

\item{Species}{A species name for the bacterial species being analyzed. This should be a text string, currently
"Salmonella" and "E.coli" are accepted as possible species for EPA-NG analysis.}
}
\value{
A phylogeny plot with the MRCA and descendants colored in red, representing the final clade of resulting hits.
}
\description{
This function will take the final calculated clade MRCA and generate a circular phylogeny with red lines representing
the path from the final clade MRCA to all of its descendants (the final hits). Because previous calculations are run
using an unrooted tree, this function recalculates the MRCA in the rooted tree and then colors and plots based on
the structure of the rooted tree for improved visualization. Note that this function will not work correctly on
the unrooted tree and it is important to not confuse the two different trees becasue of small node differences
between them. After running the phylogeny plotting function, plot the outputted object to generate the actual plot.
Because of the size of the actual phylogeny, it is reccomended  to save the resulting plot as a PNG in dimensions
of at least 15,000 x 15,000 pixels. An outside viewer should be used to open the PNG and zoom in to analyze the
resulting phylogeny figure. Larger dimensions will increase clarity and allow better zooming at the cost of larger
image size.
}
\examples{
Plot <- Phylogeny_Plotting(MRCA = 1352, Species = "Salmonella")
}
