% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sero_Placement_Functions.R
\name{Placement_Node_Selector}
\alias{Placement_Node_Selector}
\title{Distal / Proximal Node Selection}
\usage{
Placement_Node_Selector(Distal_Node, Distal_Length, Tree, Root, Dist_Table)
}
\arguments{
\item{Distal_Node}{An integer value for the distal node resulting from a placement.}

\item{Distal_Length}{An integer value for the distal length between the node inserted by the placement algorithm
and the distal node.}

\item{Tree}{A phylogenetic tree representing the reference sequences in Newick format. This tree is provided as
part of the Seroplacer package.}

\item{Root}{An integer value for the root node of the phylogenetic tree.}

\item{Dist_Table}{A table of pairwise distances between all pairs of nodes in the phylogenetic tree. Calculated from
dist.nodes {ape}.}
}
\value{
An integer value representing either the distal or proximal node adjacent to the node inserted by the
placement algorthim
}
\description{
This is a function built into the MRCA calculation function which performs the initial determination of whether
each placement insertion is closer to the distal or proximal node (Step 1. from Clade_Hit_Finder_Pendant_Final).
}
\examples{
Missing
}
