% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sero_Placement_Functions.R
\name{Placement_Results_Output_Full}
\alias{Placement_Results_Output_Full}
\title{Serovar Prediction Results
This function generates the output of placement results from the final resulting clade calculated using the
MRCA calculation function (Clade_Hit_Finder_Pendant_Final). The MRCA calculation function determines the final
clade MRCA, and this function will analyze the resulting hits (all descendants of that MRCA) and report the
resulting serovars in a data table. This version of the results generation will not trim any serovars with low
representation. This will report all serovars for any hits found below the final pendant adjusted MRCA.}
\usage{
Placement_Results_Output_Full(MRCA, Species)
}
\arguments{
\item{MRCA}{An integer value for the node which is the final pendant adjusted MRCA.}

\item{Species}{A species name for the bacterial species being analyzed. This should be a text string, currently
"Salmonella" and "E.coli" are accepted as possible species for EPA-NG analysis.}
}
\value{
#A data table containing the following information for each resulting serovar:
1. The number of matches in the final resulting clade. 2. The fraction of total matches which that serovar
represents. 3. The maximum depth, or the distance from the MRCA to the farthest hit in the clade (calculated
for each serovar). 4. The sum of branch lengths, or the total sum of branch lengths from each hit to the MRCA
(calculated for each serovar).
}
\description{
Serovar Prediction Results
This function generates the output of placement results from the final resulting clade calculated using the
MRCA calculation function (Clade_Hit_Finder_Pendant_Final). The MRCA calculation function determines the final
clade MRCA, and this function will analyze the resulting hits (all descendants of that MRCA) and report the
resulting serovars in a data table. This version of the results generation will not trim any serovars with low
representation. This will report all serovars for any hits found below the final pendant adjusted MRCA.
}
\examples{
Serovar_Results_Table_Full <- Placement_Results_Output(MRCA = 1352, Species = "Salmonella")
}
