% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sero_Placement_Functions.R
\name{Sero_Result}
\alias{Sero_Result}
\title{Simple Serovar Prediction}
\usage{
Sero_Result(Sero)
}
\arguments{
\item{Sero}{A serovar results table which is the output of Placement_Results_Output.}
}
\value{
A serovar prediction if a serovar represents at least 30% of the resulting hits, otherwise no result.
}
\description{
This function will report a predicted serovar from Placement_Results_Output. This function requires that a serovar shows at least
30% representation to report, and will report all serovars over the 30% representation threshold. If there are no serovars with
at least 30% presence then it will report that no serovar match was found. It is highly reccomended to analyze the results table
in depth to understand the accuracy of the serovar prediction, as the table will contain a complete visualization of all results.
}
\examples{
Missing
}
