% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sero_Placement_Functions.R
\name{string.diff}
\alias{string.diff}
\title{String Nucleotide Mismatch Calculation}
\usage{
string.diff(a, b, exclude = c("n", "N", "?"), ignore.case = TRUE)
}
\arguments{
\item{a}{A character string, equal in length to b.}

\item{b}{A character string, equal in length to a.}

\item{exclude}{Characters which to exclude from mismatch calculation.}

\item{ignore.case}{TRUE/FALSE whether to ignore case of characters.}
}
\value{
A number value which is the number of nucleotide differences between the two inputted strings.
}
\description{
This function calculates the number of nucleotide position differences between two equal length character strings.
We use this in order to calculate all possible mismatch values between query and references sequences in order
to re-sort the reference sequences into the appropriate order.
}
\examples{
string.diff(a = "AAAAACA",b = "AAAAGCA",exclude = c("n","N","?"),ignore.case = TRUE)
}
