% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sero_Placement_Functions.R
\name{Super_HAM_Combinator}
\alias{Super_HAM_Combinator}
\title{Combination Testing}
\usage{
Super_HAM_Combinator(Hamming_Table, Combination)
}
\arguments{
\item{Hamming_Table}{A 7x7 table containing the number of nucleotide mismatches between 7 query and 7 reference alleles.}

\item{Combination}{A vector of the numbers 1-7 in a specific combination order.}
}
\value{
An integer value which is the  sum of nucleotide mismatches for the specific 1-7 combination which was
tested. This function is applied over a table of all possible 5040 combinations to test all possible allele orders.
}
\description{
A function for calculating the number of mismatches in each possible combination of allele pairings between 7 query
alleles and 7 reference alleles. A 7x7 table is calculated with the number of nucleotide mismatches between each
query and reference initially. The combination function then calculates the sum of nucleotide mismatches for each
possible set of pairings of 7 alleles. There are a total of 7! = 5040 combinations which are tested to determine
which sum is the lowest, representing the best possible orientation of reference alleles relative to the query alleles.
}
\examples{
Missing
}
