% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sero_Placement_Functions.R
\name{Super_HAM_Placer}
\alias{Super_HAM_Placer}
\title{Reference Sequence Re-ordering}
\usage{
Super_HAM_Placer(TestSequence, ReferenceSequence)
}
\arguments{
\item{TestSequence}{A vector of nucleotide sequences containing 7 sequences representing the 7 16S query alleles.}

\item{ReferenceSequence}{A vector of nucleotide sequences containing 7 sequences representing the 7 16s reference
alleles. This vector should contain 8 total elements, the first being a name, followed by 7 sequence elements.}
}
\value{
A vector of 7 numbers (1-7) which represents the order of reference sequences which contains the lowest
total number of nucleotide mismatches when compared to the query sequence.
}
\description{
Because we are working with a multi-copy marker gene, it is important to re-sort the reference 16S sequences into
the order which best matches the user submitted query sequences. This function will calculate the number of nucleotide
differences between all query and reference sequences and then determine the combination which results in the
minimum number of mismatches summed across all allele pairings.
}
\examples{
Best_Sequence_Order <- Super_HAM_Plaer(TestSequence = Sequence1, ReferenceSequence = Sequence2)
}
